Switch devel package back to the "compat" ABI-5
The switch to from ncurses 5.9 to 6.4 for xsconsole's convenience has
been unwantingly linking binaries against both abi5 (through
libedit) and abi6 (directly).

To fix this and make sure no package gets built against abi6 again, we
change the devel package to shipping the abi5 headers and .so
symlinks.  We achieve this essentially by swapping the "make install"
order of the abi5 and abi6 configurations.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>